Create an Accessibility Guide to Mastodon
This guide would cover reviews of Android apps (those with iOS can maybe assist with the reviews for those apps as I don't have any iOS devices), and other accessibility tips such as how to draft toots, how to make a thread, and other similar topics not yet covered. It could be a working document for folks to comment on, which can be sort of like an accessibility FAQ. 

At the moment, I'm not quite done writing it. And I haven't input in my images yet. I'm covering accessibility options within preferences and the android app reviews I did. I'll come back and edit it more once I have a bit more spoons, but I wanted to make sure I saved it!